requirements.txt
airtable-python-wrapper
feedparser
requests
